+++ djangojs/settings.py
@@ -130,6 +130,7 @@ INSTALLED_APPS = (
     'django.contrib.admin',
     # Uncomment the next line to enable admin documentation:
     'djangojs',
+    'djangojs.fake',
 )
 
 # A sample logging configuration. The only tangible logging