+++ djangojs/tests.py
@@ -462,14 +462,34 @@ class StorageGlobberTest(TestCase):
         self.assertNotIn('js/test/libs/qunit.js', result)
 
 
-class SettingsTest(TestCase):
-    urls = 'djangojs.test_urls'
+@override_settings(USE_I18N=True, LANGUAGE_CODE='en', JS_I18N=['djangojs'])
+class I18nIncludeTest(TestCase):
+    def setUp(self):
+        from djangojs.urls import js_info_dict
+        self.packages = js_info_dict()['packages']
 
-    @override_settings(JS_URLS_NAMESPACES=['ns1'])
-    def test_urls_namespaces(self):
-        '''Should only include JS_URLS_NAMESPACES'''
+    def test_include_translation(self):
+        '''Should include apps listed in JS_I18N'''
+        self.assertTrue('djangojs' in self.packages)
+
+    def test_not_include_translation(self):
+        '''Should not include apps not listed in JS_I18N'''
+        self.assertFalse('djangojs.fake' in self.packages)
 
 
+@override_settings(USE_I18N=True, LANGUAGE_CODE='en', JS_I18N_EXCLUDE=['djangojs'])
+class I18nExcludeTest(TestCase):
+    def setUp(self):
+        from djangojs.urls import js_info_dict
+        self.packages = js_info_dict()['packages']
+
+    def test_include_translation(self):
+        '''Should exclude apps listed in JS_I18N_EXCLUDE'''
+        self.assertFalse('djangojs' in self.packages)
+
+    def test_not_include_translation(self):
+        '''Should not exlude apps not listed in JS_I18N_EXCLUDE'''
+        self.assertTrue('djangojs.fake' in self.packages)
 
 
 class TapAssertionTest(unittest.TestCase):